fix Jest accidentally testing .d.ts file
