--- script/pre-prepublish.js
@@ -1,11 +1,0 @@
-require('shelljs/global');
-
-const path = require('path');
-
-rm('-Rf', path.join(__dirname, '..', 'node_modules', '@types', 'react'));
-
-cp(
-  '-r',
-  path.join(__dirname, '..', 'node_modules_temp', '@types', 'react', '/'),
-  path.join(__dirname, '..', 'node_modules', '@types')
-);
